+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.helloworld",
+    package = "com.sisuratuin",
 )
 
 android_resource(
     name = "res",
-    package = "com.helloworld",
+    package = "com.sisuratuin",
     res = "src/main/res",
 )
