+++ android/app/src/main/java/com/sisuratuin/MainActivity.java
@@ -1,4 +1,4 @@
-package com.helloworld;
+package com.sisuratuin;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "HelloWorld";
+        return "SiSuratUIN";
     }
 }
