+++ android/app/src/main/java/com/sisuratuin/MainApplication.java
@@ -1,4 +1,4 @@
-package com.helloworld;
+package com.sisuratuin;
 
 import android.app.Application;
