+++ android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'HelloWorld'
+rootProject.name = 'SiSuratUIN'
 
 include ':app'
