+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "HelloWorld",
-  "displayName": "HelloWorld"
+  "name": "SiSuratUIN",
+  "displayName": "SiSuratUIN"
 }
