+++ ios/SiSuratUIN.xcodeproj/xcshareddata/xcschemes/SiSuratUIN-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "HelloWorld-tvOS.app"
-               BlueprintName = "HelloWorld-tvOS"
-               ReferencedContainer = "container:HelloWorld.xcodeproj">
+               BuildableName = "SiSuratUIN-tvOS.app"
+               BlueprintName = "SiSuratUIN-tvOS"
+               ReferencedContainer = "container:SiSuratUIN.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "HelloWorld-tvOSTests.xctest"
-               BlueprintName = "HelloWorld-tvOSTests"
-               ReferencedContainer = "container:HelloWorld.xcodeproj">
+               BuildableName = "SiSuratUIN-tvOSTests.xctest"
+               BlueprintName = "SiSuratUIN-tvOSTests"
+               ReferencedContainer = "container:SiSuratUIN.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "HelloWorld-tvOSTests.xctest"
-               BlueprintName = "HelloWorld-tvOSTests"
-               ReferencedContainer = "container:HelloWorld.xcodeproj">
+               BuildableName = "SiSuratUIN-tvOSTests.xctest"
+               BlueprintName = "SiSuratUIN-tvOSTests"
+               ReferencedContainer = "container:SiSuratUIN.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "HelloWorld-tvOS.app"
-            BlueprintName = "HelloWorld-tvOS"
-            ReferencedContainer = "container:HelloWorld.xcodeproj">
+            BuildableName = "SiSuratUIN-tvOS.app"
+            BlueprintName = "SiSuratUIN-tvOS"
+            ReferencedContainer = "container:SiSuratUIN.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "HelloWorld-tvOS.app"
-            BlueprintName = "HelloWorld-tvOS"
-            ReferencedContainer = "container:HelloWorld.xcodeproj">
+            BuildableName = "SiSuratUIN-tvOS.app"
+            BlueprintName = "SiSuratUIN-tvOS"
+            ReferencedContainer = "container:SiSuratUIN.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "HelloWorld-tvOS.app"
-            BlueprintName = "HelloWorld-tvOS"
-            ReferencedContainer = "container:HelloWorld.xcodeproj">
+            BuildableName = "SiSuratUIN-tvOS.app"
+            BlueprintName = "SiSuratUIN-tvOS"
+            ReferencedContainer = "container:SiSuratUIN.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
