+++ ios/SiSuratUINTests/SiSuratUINTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface HelloWorldTests : XCTestCase
+@interface SiSuratUINTests : XCTestCase
 
 @end
 
-@implementation HelloWorldTests
+@implementation SiSuratUINTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
