+++ package.json
@@ -1,5 +1,5 @@
 {
-	"name": "HelloWorld",
+	"name": "SiSuratUIN",
 	"version": "0.0.1",
 	"private": true,
 	"scripts": {
